% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_sex_pyramid.R
\name{age_sex_pyramid}
\alias{age_sex_pyramid}
\title{Create age-sex pyramid}
\usage{
age_sex_pyramid(data, age.grp.col, sex.col, lower.limit = NULL,
  upper.limit = NULL, split.by = NULL, col.pal = 1)
}
\arguments{
\item{data}{a data.frame}

\item{age.grp.col}{a character defining the age-group column within the data.frame}

\item{sex.col}{a character defining the sex column within the data.frame}

\item{lower.limit}{a numeric defining the lower limit for the x-axis}

\item{upper.limit}{a numeric defining the upper limit for the x-axis}

\item{split.by}{a character defining a column to facet by within the data.frame}

\item{col.pal}{a numeric defining the colour palette to use}
}
\value{
an age-sex pyramid
}
\description{
Create age-sex pyramid
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
