% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate ages in years and age groups}
\usage{
calc_age(dob, ref.date, breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, Inf))
}
\arguments{
\item{dob}{a vector of dates (should be date of births)}

\item{ref.date}{a vector of dates (should be the reference date to calculate age)}

\item{breaks}{a numeric vector defining break points for age groups (the final value must be infinite)}
}
\value{
a data.frame containing a column of ages in years and a column of age groups
}
\description{
Calculate ages in years and age groups
}
\examples{
# # set dummy data
#
# set.seed(2)
#
# data = data.frame(date.of.birth = sample(seq(as.Date('1900-01-01'), as.Date('1990-12-31'), by="day"), 30, replace = TRUE),
#                   onset.date = sample(seq(as.Date('1991-01-01'), as.Date('2010-12-31'), by="day"), 30, replace = TRUE))
#
#  # use function
#
# calc_age(dob = data$date.of.birth,
#          ref.date = data$onset.date)
#
# # append results of function onto original data
#
# cbind(data, calc_age(dob = data$date.of.birth,
#                      ref.date = data$onset.date,
#                      breaks = c(0, 10, 20, 30, 40, 50, 60, Inf)))
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
