% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agegroup.R
\name{calc_agegroup}
\alias{calc_agegroup}
\title{Calculate age groups in years}
\usage{
calc_agegroup(age, breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, Inf))
}
\arguments{
\item{age}{a numeric vector of ages}

\item{breaks}{a numeric vector defining break points for age groups (the final value must be infinite)}
}
\value{
a factor vector containing age groups in years
}
\description{
Calculate age groups in years
}
\examples{
# set dummy data

set.seed(2)

age = sample(0:120, 30, replace = TRUE)

# use function

calc_agegroup(age , breaks = c(0, 20, 40, 60, Inf))
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
