% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate records}
\usage{
deduplicate(data, id.col, date.col, vars.collapse = "all",
  episode.length = Inf, break.value = ", ")
}
\arguments{
\item{data}{a data.frame}

\item{id.col}{a character specifying the column containing ID's}

\item{date.col}{a character specifying the column containing dates}

\item{vars.collapse}{a character vector specifying the columns to collapse together if there is differing information between records}

\item{episode.length}{a numeric specifying the number of days a single episode lasts}

\item{break.value}{a character specifying the value to insert as a break between collapsed variables}
}
\value{
a numeric vector containing age in years
}
\description{
Deduplicate records
}
\examples{
# define dummy data

set.seed(7)

data = data.frame(id = c(1, 1, 2, 2, 3, 4, 5, 5, NA, NA),
                  date = sample(seq(as.Date("2019-01-01"),
                                    as.Date("2019-03-01"),
                                    by = 1), size = 10),
                  spec = c("arm", "leg", "hand", "arm", "head", "hand", "blood", "blood", "arm", "arm"),
                  lab = c("Lab A", "Lab D",  "Lab B", "Lab C", "Lab A", "Lab D", "Lab F", "Lab D", "Lab B", "Lab A"))


# apply deduplicate function

deduplicate(data,
            id.col = "id",
            date.col = "date")

deduplicate(data,
            id.col = "id",
            date.col = "date",
            vars.collapse = "lab")

deduplicate(data,
            id.col = "id",
            date.col = "date",
            vars.collapse = "lab",
            episode.length = 4)

deduplicate(data,
            id.col = "id",
            date.col = "date",
            vars.collapse = "lab",
            episode.length = 4,
            break.value = " | ")
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
