% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_variable.R
\name{explode_variable}
\alias{explode_variable}
\title{Explode a single variable into multiple variables}
\usage{
explode_variable(data, vars, break.value = ", ")
}
\arguments{
\item{data}{a data.frame}

\item{vars}{a character vector specifying the columns to 'explode' into multiple seperate columns}

\item{break.value}{a character specifying the value to insert as a break between collapsed variables}

\item{id.col}{a character specifying the column containing ID's}
}
\value{
a numeric vector containing age in years
}
\description{
Explode a single variable into multiple variables
}
\examples{
# define dummy data

data = data.frame(id = c("A", "B", "C", "D"),
                  activities = c("gym, cell 1, workshop a",
                                 "workshop a, cell 2, NA",
                                 "gym, NA, NA",
                                 "NA, NA, NA"))

# apply explode_variable function

explode_variable(data, vars = "activities")
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
