% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_colnames.R
\name{format_colnames}
\alias{format_colnames}
\title{Format column names}
\usage{
format_colnames(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame
}
\description{
Format column names
}
\examples{
set dummy data with horrible column names

data = data.frame(t1 = 1, t2 = 1, t3 = 1, t4 = 1, t5 = 1)

colnames(data) = c("Col_$\%1", "2Col2", "Column number 3....",
                   "Col___4  ", "Col!!!!!!!!!!5")

# look at data

data

# apply function

format_colnames(data)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
