% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_decimals.R
\name{n_decimals}
\alias{n_decimals}
\title{Restrict the number of decimals places for a numeric value}
\usage{
n_decimals(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{number of decimals}
}
\value{
a numeric vector
}
\description{
Restrict the number of decimals places for a numeric value
}
\examples{
x = c(1.001, 1.2323221, 1.211111, 1.00001, 1, 2, 3.1, 3.66)

n_decimals(x, 0)

n_decimals(x, 5)

n_decimals(x, 10)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
