% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_clipboard.R
\name{read_from_clipboard}
\alias{read_from_clipboard}
\title{Read clipboard data into R}
\usage{
read_from_clipboard(header = TRUE, ...)
}
\arguments{
\item{header}{a logical indicating if a header has been copied}

\item{...}{other read.table options}
}
\value{
a data.frame consisting of data copied to clipboard
}
\description{
A function to easily copy data into R from a spreadsheet.
             This should not be used within a reproducible work flow, instead
             it should only be used on an ad-hoc basis
             (note: I'm not aware of the original author of this function,
             it was taken directly from a stack overflow answer)
}
\examples{
# first copy data to clipboard (i.e. highlight data in spreadsheet and press ctrl + c)

# apply function to read data from clipboard into R

data = read_from_clipboard()
}
