% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diff.R
\name{set_diff}
\alias{set_diff}
\title{Compare the set of differences between two vectors}
\usage{
set_diff(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a list of length 3 containing (X_not_in_Y, Y_not_in_X and in_X_and_y)
}
\description{
Compare the set of differences between two vectors
}
\examples{
a = c("A", "B", "C")

b = c("C", "D", "E")

set_diff(a, b)
}
