% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_na_to.R
\name{set_na_to}
\alias{set_na_to}
\title{Set NAs to a value}
\usage{
set_na_to(data, value.to.set.na.to)
}
\arguments{
\item{data}{a data.frame}

\item{value.to.set.na.to}{a character vector specifying the value to set NA to}
}
\description{
Set NAs to a value
}
\examples{
# set dummy data

set.seed(4)

data = data.frame(dates = sample(seq(as.Date('2014-01-01'), as.Date('2016-04-01'), by="day"), 20, replace = TRUE),
                  sex = factor(c("Male", "Female", "Female", NA)),
                  conf = factor(sample(c("Confirmed", "Probable", "Possible"), 20, replace = TRUE)),
                  status = sample(c("Student", "Staff", NA), 20, replace = TRUE),
                  age = sample(c(1:20, NA), 20, replace = TRUE),
                  stringsAsFactors = FALSE)

# apply function

set_na_to(data, c("Unknown"))
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
