% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_to_na.R
\name{set_to_na}
\alias{set_to_na}
\title{Set values to NA s}
\usage{
set_to_na(data, values.to.set.to.na)
}
\arguments{
\item{data}{a data.frame}

\item{values.to.set.to.na}{a character vector specifying values to set to NA}
}
\value{
a dataframe with specified values set to NA
}
\description{
This function takes a data.frame and sets specified values to NA
}
\examples{
# set dummy data

set.seed(2)

data = data.frame(dates = sample(seq(as.Date('2014-01-01'), as.Date('2016-04-01'), by="day"), 200, replace = TRUE),
                  sex = c("Male", "Female"),
                  conf = sample(c("Confirmed", "Probable", "possible"), 200, replace = TRUE),
                  status = sample(c("Student", "Staff"), 200, replace = TRUE))

# apply function

set_to_na(data, c("Male", "Student"))
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
