% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_na.R
\name{summarise_na}
\alias{summarise_na}
\title{Summarise NA values}
\usage{
summarise_na(data, plotit = FALSE)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame
}
\description{
Summarise NA values
}
\examples{
data = data.frame(geog = c("A", "B", "C", "D", NA, "E"),
                  count = c(43, 21, 19, 44, NA, NA),
                  rate = c(1.4, 2.1, 2.0, 3.3, 4.1, 1.0))

summarise_na(data, plotit = TRUE)

summarise_na(data)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
