% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_1var.R
\name{tab_1var}
\alias{tab_1var}
\title{Tabulate 1 variable}
\usage{
tab_1var(data, var = NULL, complete = FALSE, arrange.factor.by = "value",
  show.percentage = TRUE, n.decimals = 0, include.na.percentage = TRUE,
  include.total = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{var}{a character specifying the variable to tabulate, if set to NULL the first variable will be used}

\item{complete}{a logical specifying whether to use all levels for factor variables}

\item{arrange.factor.by}{a character with value either "level" or "value" describing how a factor variable should be ordered}

\item{show.percentage}{a logical specifying whether to show percentages in output}
}
\description{
Tabulate 1 variable
}
\examples{
# set dummy data

set.seed(4)

data = data.frame(date = sample(seq(as.Date("2017-01-01"), as.Date("2018-06-01"), 1), 200, replace = TRUE),
                  sex = factor(c("M", "M", "F", NA, NA), c("F", "M", "Unk")),
                  conf = sample(c("Confirmed", "Probable", "Probable"), 200, replace = TRUE),
                  status = sample(c("Student", "Staff", NA), 200, replace = TRUE),
                  geog = sample(c("South", "North"), 200, replace = TRUE))

# apply function

tab_1var(data, "sex")

# using dplyr syntax

data \%>\%
select(sex) \%>\%
  tab_1var()
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
