% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2vars.R
\name{tab_2vars}
\alias{tab_2vars}
\title{Tabulate 2 variables}
\usage{
tab_2vars(data, row.var = NULL, col.var = NULL, complete = TRUE,
  include.row.total = TRUE, include.col.total = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{row.var}{a character specifying the variable to tabulate along rows, if set to NULL the first variable will be used}

\item{complete}{a logical specifying whether to use all levels for factor variables}

\item{include.row.total}{a logical specifying whether to include row totals}

\item{include.col.total}{a logical specifying whether to include column totals}

\item{cols.var}{a character specifying the variable to tabulate along columns, if set to NULL the second variable will be used}
}
\description{
Tabulate 2 variables
}
\examples{
# set dummy data

set.seed(4)

data = data.frame(date = sample(seq(as.Date("2017-01-01"), as.Date("2018-06-01"), 1), 200, replace = TRUE),
                  sex = factor(c("M", "M", "F", NA, NA), c("F", "M", "Unk")),
                  conf = sample(c("Confirmed", "Probable", "Probable"), 200, replace = TRUE),
                  status = sample(c("Student", "Staff", NA), 200, replace = TRUE),
                  geog = sample(c("South", "North"), 200, replace = TRUE))

# apply function

tab_2vars(data, row.var = "sex", col.var = "conf")

# using dplyr syntax

data \%>\%
select(sex, conf) \%>\%
  tab_2vars()
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
