% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_each.R
\name{tab_each}
\alias{tab_each}
\title{Tabulate each variable separately}
\usage{
tab_each(data, cols = NULL, trunc.length = NULL,
  row.break.value = ".", complete = FALSE,
  arrange.factor.by = "value", show.percentage = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{cols}{a character specifying the variables to, if set to NULL all variables will be used}

\item{trunc.length}{a numeric specificying the number of characters to use when truncating the output}

\item{row.break.value}{the break value to use to separate variables in the output}

\item{complete}{a logical specifying whether to use all levels for factor variables}

\item{arrange.factor.by}{a character with value either "level" or "value" describing how a factor variable should be ordered}

\item{show.percentage}{a logical specifying whether to show percentages in output}
}
\description{
Tabulate each variable separately
}
\examples{
# set dummy data

set.seed(4)

data = data.frame(date = sample(seq(as.Date("2017-01-01"), as.Date("2018-06-01"), 1), 200, replace = TRUE),
                  sex = factor(c("M", "M", "F", NA, NA), c("F", "M", "Unk")),
                  conf = sample(c("Confirmed", "Probable", "Probable"), 200, replace = TRUE),
                  status = sample(c("Student", "Staff", NA), 200, replace = TRUE),
                  geog = sample(c("South", "North"), 200, replace = TRUE))

# apply function

tab_each(data, cols = c("date", "sex", "conf", "geog"))

# using dplyr syntax

data \%>\%
select(date, sex, conf, geog) \%>\%
  tab_each()
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
