% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stack_values.R
\name{tab_stack_values}
\alias{tab_stack_values}
\title{Tabulate values for multiple variables and stack ontop of one another}
\usage{
tab_stack_values(data, cols = NULL, fill.value = ".")
}
\arguments{
\item{data}{a data.frame}

\item{fill.value}{a value used to represent missing values}
}
\value{
a data.frame
}
\description{
Tabulate values for multiple variables and stack ontop of one another
}
\examples{
# set dummy data

set.seed(2)

data = data.frame(id = 1:30,
                  pizza = sample(c("No", "Yes", NA), 30, replace = TRUE),
                  pasta = sample(c("No", "Yes", NA), 30, replace = TRUE),
                  salad = sample(c("No", "Yes"), 30, replace = TRUE),
                  sex = sample(c("Male", "Female", NA), 30, replace = TRUE))

# use function

tab_stack_values(data, cols = c("pizza", "pasta", "salad"))

# using dplyr syntax

data \%>\%
  select(pizza, pasta, salad) \%>\%
  tab_stack_values() 
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
