% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{Tile plot}
\usage{
tile_plot(data, x.col, y.col, x.lab = "", y.lab = "", text = FALSE,
  colour = "red", angle = 0, label.breaks = 0,
  rescale.by.row = FALSE, keep.row.order = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{x.col}{a character containing the column to be used along the x-axis of the tile plot}

\item{y.col}{a character containing the column to be used along the y-axis of the tile plot}

\item{x.lab}{a character containing the label for the x-axis}

\item{y.lab}{a character containing the label for the y-axis}

\item{text}{a logical specifying if numeric values should be overlayed as text}

\item{colour}{a character containing the colour for the tile plot}

\item{angle}{a numeric to specify the x-axis label angel for the epicurve}

\item{label.breaks}{a numeric specifying x-axis label breaks}

\item{rescale.by.row}{a logical specifying if the colours on the tile plot should be relative to all tiles or relative to other tiles within a single row}

\item{keep.row.order}{a logical specifying if the order of the rows should be kept as is or changed to most common}
}
\value{
a tile plot
}
\description{
Tile plot
}
\examples{
# create dummy data

set.seed(3)

data = data.frame(geog = sample(c("Vienna", "Vienna", "Vienna", "Vienna",
                                  "Salzburg", "Innsbruck", "Graz", "Graz",
                                  "Linz", "Klagenfurt", "Villach"), 5000, replace = TRUE),
                  age.group = sample(c("0-19", "20-39", "40-59", "60+", "60+"), 5000, replace = TRUE),
                  week = factor(sample(paste0("week.", 1:30), 5000, replace = TRUE),
                                levels = paste0("week.", 1:30)))

# example plot 1

tile_plot(data, x.col = "week", y.col = "geog", keep.row.order = TRUE)

# example plot 2

tile_plot(data, x.col = "week", y.col = "geog", keep.row.order = FALSE)

# example plot 3

tile_plot(data, x.col = "week", y.col = "geog", text = TRUE)

# example plot 4

tile_plot(data, x.col = "week", y.col = "geog", text = TRUE, label.breaks = 2)

# example plot 5

tile_plot(data, x.col = "week", y.col = "geog", text = TRUE, label.breaks = 2,
          rescale.by.row = TRUE)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
