% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{Transpose a data.frame}
\usage{
trans(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data.frame
}
\description{
Transpose a data.frame
}
\examples{
data = data.frame(geog = c("A", "B", "C", "D"),
                  count = c(43, 21, 19, 44),
                  rate = c(1.4, 2.1, 2.0, 3.3))

trans(data)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
