% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_clipboard.R
\name{write_to_clipboard}
\alias{write_to_clipboard}
\title{Write R data to clipboard}
\usage{
write_to_clipboard(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{a data.frame to be copied to clipboard}

\item{row.names}{a logical indicating if row names should be included}

\item{col.names}{a logical indicating if column names should be included}

\item{...}{other read.table options}
}
\value{
data.frame copied to clipboard
}
\description{
A function to easily copy data out of R into a spreadsheet.
             This should not be used within a reproducible work flow, instead
             it should only be used on an ad-hoc basis
             (note: I'm not aware of the original author of this function,
             it was taken directly from a stack overflow answer)
}
\examples{
# set dummy data

set.seed(2)

data = data.frame(date = sample(seq(as.Date("2017-03-01"), as.Date("2017-08-01"), 1), 20, replace = TRUE),
                  sex = sample(c("Female", "Male"), 20, replace = TRUE))

# apply function to write data to clipboard

write_to_clipboard(data)

# now paste data into a spreadsheet (i.e. open a spreadsheet and press ctrl + v)
}
