% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age in years}
\usage{
calc_age(dob, ref.date)
}
\arguments{
\item{dob}{a vector of dates (should be date of births)}

\item{ref.date}{a vector of dates (should be the reference date to calculate age)}
}
\value{
a numeric vector containing age in years
}
\description{
Calculate age in years
}
\examples{
# set dummy data

# set.seed(2)

data = data.frame(date.of.birth = sample(seq(as.Date('1900-01-01'), as.Date('1990-12-31'), by="day"), 30, replace = TRUE),
                  onset.date = sample(seq(as.Date('1991-01-01'), as.Date('2010-12-31'), by="day"), 30, replace = TRUE))

# use function

calc_age(dob = data$date.of.birth,
         ref.date = data$onset.date)
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
