% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{epicurve}
\alias{epicurve}
\title{Create epicurve}
\usage{
epicurve(x, date.col, time.period, fill.by = NULL, split.by = NULL,
  shade.by = NULL, start.at = NULL, stop.at = NULL, xlab = NULL,
  ylab = "Number of cases", fill.by.legend.title = NULL,
  shade.by.legend.title = NULL, angle = 0, col.pal = 1,
  label.breaks = 0, epi.squares = TRUE, blank.background = FALSE,
  na.rm = FALSE)
}
\arguments{
\item{x}{a data.frame}

\item{date.col}{a character specifying the column containing dates}

\item{time.period}{a character of the desired time period for the epicurve, this can either be day, year, month, year.month, iso.year, iso.week, iso.year.week, use.date.col.as.is NOTE: if time.period = use.date.col.as.is the date.col will be treated as a factor and will be used as the x-axis}

\item{fill.by}{a character specifying the column to stratify by using colour}

\item{split.by}{a character specifying the column to facet by}

\item{shade.by}{a character specifying the column to stratify by using shading}

\item{start.at}{a character in the format "yyyy-mm-dd" to specify the start date for the epicurve}

\item{stop.at}{a character in the format "yyyy-mm-dd" to specify the stop date for the epicurve}

\item{xlab}{a character specifying the x-axis label for the epicurve}

\item{ylab}{a character specifying the y-axis label for the epicurve}

\item{fill.by.legend.title}{a character specifying the fill legend title}

\item{shade.by.legend.title}{a character specifying the shade legend title}

\item{angle}{a numeric to specify the x-axis label angel for the epicurve}

\item{col.pal}{a numeric specifying the colour palette  (range 1-8 inclusive) OR a character stating 'phe' to use the phe colour palette}

\item{label.breaks}{a numeric specifying the interval for x-axis label breaks}

\item{epi.squares}{a logical  specifying if episquares should be used on the epicurve}

\item{blank.background}{a logical  specifying if the figure background should be blank}

\item{na.rm}{a logical  specifying if missing dates should be removed from the epicurve}
}
\value{
an epicurve
}
\description{
Create epicurve
}
\examples{
# set dummy data

set.seed(2)

data = data.frame(dates = sample(seq(as.Date('2014-01-01'), as.Date('2016-04-01'), by="day"), 200, replace = TRUE),
                  sex = c("Male", "Female"),
                  conf = sample(c("Confirmed", "Probable", "possible"), 200, replace = TRUE),
                  status = sample(c("Student", "Staff"), 200, replace = TRUE))

data$dates.year.month = factor(format(data$dates, "\%Y_\%m"),
                               levels = unique(format(seq(min(data$dates),
                                                          max(data$dates), 1), "\%Y_\%m")))

# use function

epicurve(data, date.col = "dates", time.period = "iso.year.week",
         fill.by="sex", split.by=NULL, shade.by=NULL,
         start.at = "2014-08-01", stop.at = "2015-06-28",
         xlab="Year week", ylab="Count",
         fill.by.legend.title = "Sex", shade.by.legend.title = NULL, angle=90,
         col.pal="phe", label.breaks = 1, epi.squares = TRUE, blank.background = TRUE,
         na.rm = TRUE)

epicurve(x = data, date.col = "dates.year.month", time.period = "use.date.col.as.is",
         fill.by="status", split.by=NULL, shade.by=NULL,
         start.at = "2015-02-01", stop.at = "2015-06-22",
         xlab="Year_week", ylab="Count",
         fill.by.legend.title = "", shade.by.legend.title = NULL, angle=90,
         col.pal=1, label.breaks = 0, epi.squares = FALSE, na.rm = TRUE)

epicurve(data, date.col = "dates", time.period = "year.quarter",
         fill.by="sex", split.by=NULL, shade.by=NULL,
         start.at = "2014-01-01", stop.at = "2015-06-22",
        xlab=NULL, ylab="Count",
         fill.by.legend.title = NULL, shade.by.legend.title = NULL, angle=90,
         col.pal=8, label.breaks = 0, epi.squares = FALSE, na.rm = TRUE)

epicurve(data, date.col = "dates", time.period = "year.month",
         fill.by="sex", split.by=NULL, shade.by=NULL,
         start.at = "2014-01-01", stop.at = "2016-01-31",
         xlab=NULL, ylab="Count",
         fill.by.legend.title = "Sex", shade.by.legend.title = NULL, angle=90,
         col.pal=2, label.breaks = 0, epi.squares = FALSE, na.rm = FALSE)

epicurve(data, date.col = "dates", time.period = "day",
         fill.by="sex", split.by=NULL, shade.by=NULL,
         start.at = "2014-01-01", stop.at = "2014-02-19",
         xlab=NULL, ylab="Count", fill.by.legend.title = NULL, angle=90,
         col.pal=4, label.breaks = 1, epi.squares = FALSE, na.rm = TRUE)

epicurve(data, date.col = "dates.year.month", time.period = "use.date.col.as.is",
         fill.by="sex", split.by=NULL, shade.by=NULL,
         start.at = "2014-01-01", stop.at = "2014-02-19",
         xlab=NULL, ylab="Count", fill.by.legend.title = NULL, angle=90,
         col.pal=4, label.breaks = 0, epi.squares = FALSE, na.rm = TRUE)

epicurve(data, date.col = "dates", time.period = "month",
         fill.by="sex", split.by=NULL, shade.by="conf",
         start.at = "2014-01-01", stop.at = "2016-04-20",
         xlab=NULL, ylab="Count",
         fill.by.legend.title = NULL, shade.by.legend.title = NULL, angle=0,
         col.pal=6, label.breaks = 0, epi.squares = FALSE, na.rm = TRUE)

epicurve(data, date.col = "dates", time.period = "month",
         fill.by="conf", split.by="conf", shade.by="sex",
         start.at = "2014-01-01", stop.at = "2016-04-20",
         xlab=NULL, ylab="Count",
         fill.by.legend.title = "Confirmed status", shade.by.legend.title = "Sex",
         angle=0, col.pal="phe", label.breaks = 0, epi.squares = TRUE, na.rm = TRUE)

epicurve(data, date.col = "dates", time.period = "month",
         fill.by="sex", split.by="conf", shade.by=NULL,
        start.at = "2014-01-01", stop.at = "2016-04-20",
         xlab="Month", ylab="Count",
         fill.by.legend.title = NULL, shade.by.legend.title = NULL, angle=0,
         col.pal=7, label.breaks = 0, epi.squares = TRUE, na.rm = TRUE)
         
}
\author{
Daniel Gardiner (daniel.gardiner@phe.gov.uk)
}
